moved modelDefaults() and pal to model_init.cpp
